requirements.txt
Flask_Bootstrap==3.3.7.1
Werkzeug==2.0.1
opencv_python==4.5.3.56
Flask==2.0.1
pytesseract==0.3.8
numpy==1.21.2
dominate==2.6.0
pandas==1.3.2
PyMuPDF==1.18.17
Pillow==8.3.2
